% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{BalancedSample}
\alias{BalancedSample}
\title{Obtain a class-balanced sample}
\usage{
BalancedSample(df, sample_size = 5000, class_col = class_col)
}
\arguments{
\item{df}{An object of class \code{data.frame}.}

\item{sample_size}{Numeric. Size of the sample from each class.}

\item{class_col}{A character vector with the name of the column that identify 
the classes.}
}
\value{
Returns a \code{data.frame} with n=\code{sample_size} rows per class
        randomly sampled.
}
\description{
This function samples the rows of a \code{data.frame} with balanced classes.
Useful when original training set has severely unbalanced classes.
}
