% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_beads_dataset.R
\name{Beads_TrainTest}
\alias{Beads_TrainTest}
\title{Wrapper function to generate datasets for model training}
\usage{
Beads_TrainTest(
  sample_size = 30,
  method = method,
  bsample = 5000,
  class_col = "BeadsSmp_ID",
  ...
)
}
\arguments{
\item{sample_size}{A numeric vector. Number of samples (files) to be used.}

\item{method}{A character vector. Clustering method for \code{BeadsUnsup}.}

\item{bsample}{A numeric vector. Size of the sample from each class for 
to be passed to \code{BalancedSample} function.}

\item{class_col}{A character vector with to label the column that identify 
the classes.}
}
\description{
This function generates training and test data sets for bead classification
model training. First it uses \code{BeadsUnsup} to select for 'good' samples.
Then it uses \code{BalancedSample} and \code{TrainTest} to generate a
class-balanced training and test sets.
}
