% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_beads_dataset.R
\name{BeadsIDmerge}
\alias{BeadsIDmerge}
\title{Merge beads-only and cells-only}
\usage{
BeadsIDmerge()
}
\description{
This function labels and merges beads-only and cells-only files collected 
on the same date separately. Column names in these files should be edited as 
needed to ensure are the same across files prior using this function. 
Working directory should be the folder that contain the files of interest. 
Files corresponding to the same date will be merged and new FCS files with 
a class column will be outputted. After the merge, files can be normalized
using a bead normalization software.
}
