% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_beads_dataset.R
\name{BeadsUnsup}
\alias{BeadsUnsup}
\title{Beads unsupervised classification}
\usage{
BeadsUnsup(
  df,
  method = c("GMM", "k_means"),
  beads_ch = beads_ch,
  n_clusters = 2,
  CV_max = 0.05,
  filename = filename,
  output_path = output_path,
  FCSdesc = FCSdesc
)
}
\arguments{
\item{df}{An object of class \code{data.frame}.}

\item{method}{A character vector with the name of the clustering algorithm 
to be used. Options are Gaussian Mixture Models 'GMM' or K-means
'k_means'.}

\item{beads_ch}{A numeric vector with the column indexes for the beads channels.}

\item{n_clusters}{A numeric vector indicating the number of clusters to be targeted.}

\item{CV_max}{A numeric vector indicating the maximum value for the coefficient 
of variation (CV) to evaluate beads cluster.}

\item{filename}{A character vector with the name of the file that originated
the data.frame.}

\item{output_path}{A character vector with full path names.}

\item{FCSdesc}{FCS column descriptions (extracted from \code{flowFrame} object)}
}
\value{
If CV< \code{CV_max} returns a \code{data.frame} with new class column
        for beads or cells. Otherwise, returns a vector = 0.
}
\description{
This function uses clustering to classify the events into beads or cells. 
Beads are expected to be a group of very little dispersion, so the 
classification is assessed by looking into the Coefficient of Variation of
the rows classified as beads. If this is > \code{CV_max}, the classification 
failed and the \code{df} is converted to 0. Otherwise, \code{df} with a new
class column is outputted. The summary stats are written in CSV files either
case.
}
