% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{ClassVarToFactor}
\alias{ClassVarToFactor}
\title{Converts the class labels from numeric to factor.}
\usage{
ClassVarToFactor(df, class_col = class_col, name_0 = name_0, name_1 = name_1)
}
\arguments{
\item{df}{An object of class \code{data.frame}.}

\item{class_col}{A character vector with the name of the column that identify 
the classes.}

\item{name_0}{A character vector with the name for the class==0}

\item{name_1}{A character vector with the name for the class==1}
}
\value{
Returns a \code{data.frame} with the original class labels converted
        into factors.
}
\description{
A function to make the class column compatible with \code{caret} training function
for classification. Should be used when classes are labeled as numeric and
need to be transform to a factor.
}
