% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_utils.R
\name{extract_data}
\alias{extract_data}
\title{Extract data (fcs, csv or txt)}
\usage{
extract_data(
  file_type = ".fcs|.FCS",
  sampling = FALSE,
  sample_size = 10,
  SpecifyList = FALSE,
  f_list
)
}
\arguments{
\item{file_type}{A regular expression for the file extension. Only file names 
which match the regular expression will be returned. Data types suported
are .FCS, .CSV and .TXT}

\item{sampling}{Logical. Should a sample of size \code{sample_size} extracted 
from the list of files? If \code{FALSE} the function will extract all 
files of the list. Default is \code{FALSE}}

\item{sample_size}{Numeric. If \code{sampling=TRUE} this would be the sample 
size.}

\item{SpecifyList}{Logical. If \code{TRUE}, a vector of file names shold be 
given in \code{f_list}. Default is \code{FALSE}}

\item{f_list}{If \code{SpecifyList=TRUE}, this should be a vector of file names.}
}
\value{
Returns a \code{list} of \code{data.frames} of the extracted data.
}
\description{
This function reads files of current folder into a \code{list} of 
\code{data.frames}
}
