% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_debris_dataset.R
\name{post_gate}
\alias{post_gate}
\title{Post-gate debris labeling}
\usage{
post_gate(bsample = 5000, path_pregated = "../")
}
\arguments{
\item{bsample}{A numeric vector. Size of the sample from each class for 
to be passed to \code{BalancedSample} function.}

\item{path_pregated}{A character vector with full path names. Default is the 
one directory up.}
}
\description{
This function compares the pre-gated files with the gated files and adds the
label of noise and cells to the events in each file. Then it uses 
\code{BalancedSample} and \code{TrainTest} to generate a class-balanced 
training and test sets. The working directory should be the folder that 
contains the gated files and the pre-gated files should be one directory up.
}
