% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_debris_dataset.R
\name{pre_gate}
\alias{pre_gate}
\title{Preprocess data before debris manual gating}
\usage{
pre_gate(
  sample_size = 20,
  use.current.model = TRUE,
  model_beads = model_beads,
  alg_bd = "RF"
)
}
\arguments{
\item{sample_size}{A numeric vector. Number of samples (files) to be used.}

\item{use.current.model}{Logical. If \code{FALSE}, an object of class \code{train} can 
be passed to the argument \code{model_beads}. Default is \code{TRUE}.}

\item{model_beads}{Object of class \code{train}. Needed if \code{use.current.model} 
is \code{FALSE}.}

\item{alg_bd}{A character vector with the name of the algorithm used to train
\code{model_beads}. It can be 'RF' for Random Forest or 'XGB' for XGBoost.}
}
\description{
This function selects a random sample and reads the FCS files, remove zeros 
and beads and then adds a row ID column and writes a new FCS file. It uses 
files on current working directory. Files should have same parameters 
desc/column names.
}
