% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{predict_cl}
\alias{predict_cl}
\title{Predict class}
\usage{
predict_cl(
  df,
  model = model,
  alg = c("RF", "XGB"),
  features = features,
  label = label
)
}
\arguments{
\item{df}{An object of class \code{data.frame}.}

\item{model}{Object of class \code{train} that contains the model to be used 
for classification of the new data.}

\item{alg}{A character vector with the name of the classification algorithm 
used to train \code{model}.}

\item{features}{A character vector with the features used to train \code{model}
(also known as predictors).}

\item{label}{A character vector with the name of the positive class. (e.g. 'beads')}
}
\value{
Returns a numeric vector with the row indexes of the positive class.
}
\description{
This function predicts the class of the events (rows) based on a classification 
model previously trained and returns the row indexes of the positive class.
}
