% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_noise.R
\name{rm_noise}
\alias{rm_noise}
\title{Noise removal function}
\usage{
rm_noise(
  file_type = ".fcs|.FCS",
  rm_beads = TRUE,
  rm_debris = TRUE,
  use.current.model.beads = TRUE,
  use.current.model.debris = TRUE,
  model_beads = model_beads,
  model_debris = model_debris,
  alg_db = "RF",
  alg_bd = "RF"
)
}
\arguments{
\item{file_type}{A regular expression for the file extension. Only file names 
which match the regular expression will be returned. FCS only.}

\item{rm_beads}{Logical. If \code{TRUE} beads are removed. Default is \code{TRUE}.}

\item{rm_debris}{Logical. If \code{TRUE} debris is removed. Default is \code{TRUE}.}

\item{use.current.model.beads}{Logical. If \code{FALSE}, an object of class \code{train} can 
be passed to the argument \code{model_beads}. Default is \code{TRUE}.}

\item{use.current.model.debris}{Logical. If \code{FALSE}, an object of class \code{train} can 
be passed to the argument \code{model_beads}. Default is \code{TRUE}.}

\item{model_beads}{Object of class \code{train}. Needed if \code{use.current.model.beads} 
is \code{FALSE}.}

\item{model_debris}{Object of class \code{train}. Needed if \code{use.current.model.debris} 
is \code{FALSE}.}

\item{alg_db}{A character vector with the name of the algorithm used to train
\code{model_beads}. It can be 'RF' for Random Forest or 'XGB' for XGBoost.}

\item{alg_bd}{A character vector with the name of the algorithm used to train
\code{model_beads}. It can be 'RF' for Random Forest or 'XGB' for XGBoost.}
}
\description{
This function
}
