% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{rm_zeros}
\alias{rm_zeros}
\title{Remove zeros}
\usage{
rm_zeros(df, mand_idx, opt_idx, ...)
}
\arguments{
\item{df}{An object of class \code{data.frame}.}

\item{opt_idx}{A vector with the column indexes for the optional markers.}

\item{mand_inx}{A vector with the column indexes for the mandadory markers.}
}
\value{
Returns a \code{data.frame} with the invalid events already removed.
}
\description{
This function removes events with 0 expression of mandatory makers (e.g. DNA 
intercalator) as well as cells with 0 expression in all optional markers 
(e.g. CD4, CD8, EpCAM, CD31, etc.). Mandatory markers are those which should 
be expressed for an event to be considered valid. Optional markers are those 
that allow cell type identification, therefore their absolute abscence make 
the event invalid because it cannot be identified as a specific cell type.
}
