% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_utils.R
\name{t_asinh}
\alias{t_asinh}
\title{Arcsinh transformation}
\usage{
t_asinh(df, cofactor = 5)
}
\arguments{
\item{df}{An object of class \code{data.frame} containing columns to be 
scaled. The function ignores non-numeric columns.}

\item{cofactor}{The cofactor for the scaling. This could be any number >0:Inf.}
}
\value{
Returns a \code{data.frame} with the scaled data.
}
\description{
This function takes a data.frame and applies basic \code{asinh} function 
using a cofactor.
}
