% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{TrainModel}
\alias{TrainModel}
\title{Tunes and trains a classification model}
\usage{
TrainModel(
  TrainSet,
  TestSet,
  alg = c("all", "RF", "XGB"),
  class_col = class_col,
  seed = 40,
  name_0 = name_0,
  name_1 = name_1,
  label = label,
  allowParallel = TRUE,
  free_cores = 4
)
}
\arguments{
\item{TrainSet}{A \code{data.frame} with balanced classes.}

\item{TestSet}{A \code{data.frame}.}

\item{alg}{A character vector with the name of the classification algorithm 
to be used. Options are Random Forest 'RF', XGBoost 'XGB' or both 'all'.}

\item{class_col}{A character vector with the name of the column that identify 
the classes.}

\item{seed}{A numeric vector to set a seed for reproducible results.}

\item{name_0}{A character vector with the name for the class==0}

\item{name_1}{A character vector with the name for the class==1}

\item{label}{A character vector with the prefix of the RData file to be outputted.}

\item{allowParallel}{Logical. If \code{TRUE} allows parallel computation.}

\item{free_cores}{A numeric vector with the number of cores to be left free 
if \code{allowParallel=TRUE}'.}
}
\description{
This function tunes and train a model (or models) for classification. Current
version supports Random Forest and XGBoost algorithms. Allows parallelization.
The final model, feature importance, prediction results of the test dataset 
and confusion matrix are outputted as an RData object.
}
