% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{TrainSetList}
\alias{TrainSetList}
\title{Get a list of dataframes for training}
\usage{
TrainSetList(dt_ls, s_train = 0.75)
}
\arguments{
\item{dt_ls}{A \code{list} of \code{data.frame}s}

\item{s_train}{Numeric. The size of the training set. Any value between (0:1)}
}
\value{
A random subset of \code{data.frame}s as a \code{list}.
}
\description{
This function takes a list of \code{data.frame} and returns a random sample 
of the original list to be used for for training a model. The remaining 
\code{data.frame} should be used to test the model.
}
