% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_building.R
\name{TrainTest}
\alias{TrainTest}
\title{Generates datasets for model training}
\usage{
TrainTest(
  dt_ls,
  output_path = output_path,
  label = label,
  class_col = class_col
)
}
\arguments{
\item{dt_ls}{A \code{list} of \code{data.frame}s with balanced classes.}

\item{output_path}{A character vector with full path names.}

\item{label}{A character vector with the prefix of the RData file to be outputted.}

\item{class_col}{A character vector with the name of the column that identify 
the classes.}
}
\description{
This function generates training and test sets for model training with 
balanced classes using functions \code{BalancedSample} and \code{TrainSetList}.
The training and test datasets as well as vectors that identify the file of 
origin of each row are saved as an RData object.
}
